BUG: Fix VTK build using vtkExternalModuleMacros
This commit fixes the configuration and build of VTKWikiExamples as both a
VTK remote module and a stand-alone project. It ensure **first and follow-up
configurations** produce similar results and behave **identically**.

The build was broken following the introduction of
Kitware/VTK@e33b90d (ExternalModule: Improve target export to
allow build against external module).

The commit addresses the issue by introducing the WikiModuleConfig CMake
module responsible for:

* ensure module information are exported into `VTK_MODULES_DIR` by using
`vtk_module_export_info`.

* use `vtk_module_config` to ensure that variables like `VTK_LIBRARIES`
are set in the current scope when the project is build.


Former-commit-id: 766a0e16b3f4699060a33fcce46e36cfe3a53b18